functions/requirements.txt
firebase-admin==6.4.0
firebase-functions==0.1.0
ffmpeg-python==0.2.0
urllib3==2.0.7
requests==2.31.0
